That is a partial fix of issue YTDB-333, the implementation of a new
algorithm for incremental backups.
The detailed design document can be read there
https://youtrack.jetbrains.com/articles/YTDB-A-10/Incremental-backup .
Key points of this change:
1. Incremental backup is made durable by keeping a single backup file's
hash code inside the metadata.
2. The last backup file is checked during the next iteration of the
incremental backup. If the backup is broken, the file is removed, and a
backup is performed since the changes of the previous backup.
3. It is possible to provide a virtual presentation of the file system
by providing lambda functions that implement the required file system
functionality.
4. New methods for public API were introduced:
`YTDBGraphGraphTraversalSource#backup`,
`YTDBGraphGraphTraversalSource#fullBackup` and `YouTrackDB#restore`.

+++ lucene/src/main/java/com/jetbrains/youtrackdb/internal/spatial/shape/legacy/PointLegecyBuilder.java
@@ -26,10 +26,10 @@ public class PointLegecyBuilder implements ShapeBuilderLegacy<Point> {
 
   @Override
   public Point makeShape(DatabaseSessionInternal session, CompositeKey key, SpatialContext ctx) {
-    var lat = ((Double) PropertyTypeInternal.convert(session, key.getKeys().get(0),
-        Double.class)).doubleValue();
-    var lng = ((Double) PropertyTypeInternal.convert(session, key.getKeys().get(1),
-        Double.class)).doubleValue();
+    var lat = PropertyTypeInternal.convert(session, key.getKeys().get(0),
+        Double.class).doubleValue();
+    var lng = PropertyTypeInternal.convert(session, key.getKeys().get(1),
+        Double.class).doubleValue();
     return ctx.makePoint(lng, lat);
   }
